BinningBVHBuilder::test: Building a few mesh BVHs even when comprehensive_tests is false. Added SetFlushDenormsMode: restoring flush-denorm mode after doing SBVH and binning BVH builds.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@14206 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ physics/EmbreeBVHBuilder.cpp
@@ -149,7 +149,7 @@ static void splitPrimitiveFunction(
 	assert(primitive->geomID == 0);
 
 	EmbreeBVHBuilder* builder = (EmbreeBVHBuilder*)userPtr;
-	assert(primitive->primID < builder->m_num_objects);
+	assert((int)primitive->primID < builder->m_num_objects);
 	const BVHBuilderTri& tri = builder->triangles[primitive->primID];
 
 	const Vec4f prim_min = loadVec4f(&primitive->lower_x);
